% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw3dBubblePlot.R
\name{colorMixer}
\alias{colorMixer}
\title{Mix colors by position}
\usage{
colorMixer(axisColors, position)
}
\arguments{
\item{axisColors}{A length 3 character vector containing the colors of the axes.}

\item{position}{A length 3 numeric vector containing the position of the bubble}
}
\value{
A mixed color in hex
}
\description{
This function return a mixed color by adding colors of the axes. The mixed color depends on the position of the bubble.
}
