% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw3dBubblePlot.R
\name{compute2dCoordinates}
\alias{compute2dCoordinates}
\title{Convert 3D positions into 2D positions given a transformation matrix}
\usage{
compute2dCoordinates(transformMatrix, threedPositions)
}
\arguments{
\item{transformMatrix}{A 3*3 matrix transforms 3D positions from one coordinates to another. The matrix can be obtained by calculating the inverse of standard camera coordinates.}

\item{threedPositions}{A n*3 matrix containing all the positions you want to transform}
}
\value{
A n*2 matrix containing the 2d positions converted from the 3D positions
}
\description{
This function first transforms given 3D positions by the transformation matrix. Then, they were projected onto a canvas and turned into 2D positions.
}
