% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw3dBubblePlot.R
\name{computeCameraMatrix}
\alias{computeCameraMatrix}
\title{Generate camera coordinates}
\usage{
computeCameraMatrix(cameraCoordinates, cameraPosition)
}
\arguments{
\item{cameraCoordinates}{The axis directions of camera coordinates. Must be a 3*3 matrix. Rows of the matrix represent x, y, z axis direction of the camera coordinates}

\item{cameraPosition}{The position the camera located. Must be a numeric vector of length 3.}
}
\value{
a standard matrix of camera coordinates
}
\description{
This function create a standard matrix of camera coordinates
}
