% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw3dBubblePlot.R
\name{plotAxis}
\alias{plotAxis}
\title{Plot axis}
\usage{
plotAxis(axisPositions, axisColors, axisLabels)
}
\arguments{
\item{axisPositions}{A 4*3 matrix containing absolute positions of the origin, x axis end, y axis end and z axis end}

\item{axisColors}{A length 3 character vector containing three colors. The colors are represented by hex such as "#ffffff"}

\item{axisLabels}{A length 3 character vector containing the label of the three axis}
}
\description{
Plot axis
}
