% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw3dBubblePlot.R
\name{plotBubbles}
\alias{plotBubbles}
\title{Plot Bubbles}
\usage{
plotBubbles(bubblePositions, bubbleSize, bubble3DPositions, axisColors,
  bubbleLabels)
}
\arguments{
\item{bubblePositions}{A n*2 matrix containing the 2D positions of the bubbles}

\item{bubbleSize}{A numeric vector containing the size of bubbles}

\item{bubble3DPositions}{A n*3 matrix containing the 3D positions of the bubbles}

\item{axisColors}{A length 3 character vector containing colors of the axes. Use these colors to add up the color of the bubble.}

\item{bubbleLabels}{A character vector containing the labels of each bubble}
}
\description{
Plot Bubbles
}
